+++ src/sched.h
@@ -32,8 +32,8 @@ void sched_timer_kick(void);
 uint8_t sched_is_shutdown(void);
 uint16_t sched_shutdown_reason(void);
 void sched_clear_shutdown(void);
-void sched_try_shutdown(uint16_t reason);
-void sched_shutdown(uint16_t reason) __noreturn;
+void sched_try_shutdown(unsigned int reason);
+void sched_shutdown(unsigned int reason) __noreturn;
 void sched_main(void);
 
 // Compiler glue for DECL_X macros above.
